fixed compiler warning from non-debug library usage
- fixed a couple other compiler warnings

+++ src/RocketEngine/render/Mesh.h
@@ -91,7 +91,7 @@ class Mesh
 				else if (name == "texture_height")
 					number = std::to_string(heightNr++); // transfer unsigned int to stream
 
-				shader->setFloat(name.append(number), i); // now set the sampler to the correct texture unit
+				shader->setFloat(name.append(number), (float)i); // now set the sampler to the correct texture unit
 				glBindTexture(GL_TEXTURE_2D, mMeshData.textures[i].id); // and finally bind the texture
 			}
